+++ src/AppContainer.tsx
@@ -9,8 +9,8 @@ import { CompleteRegistration } from './views/CompleteRegistration';
 import { LoginGate } from './views/LoginGate';
 import { ManageAccount } from './views/ManageAccount';
 import { Signup } from './views/Signup';
-import { Testimonials } from './views/testimonials/Testimonials';
-import { TestimonialTextFull } from './views/testimonials/TestimonialTextFull';
+import { Tributes } from './views/tributes/Tributes';
+import { TributeTextFull } from './views/tributes/TributeTextFull';
 import { Welcome } from './views/Welcome';
 
 const router = createBrowserRouter([
@@ -52,12 +52,12 @@ const router = createBrowserRouter([
         element: <AboutLibrary />,
       },
       {
-        path: '/testimonials',
-        element: <Testimonials />,
+        path: '/tributes',
+        element: <Tributes />,
       },
       {
-        path: '/testimonials/:id',
-        element: <TestimonialTextFull />,
+        path: '/tributes/:id',
+        element: <TributeTextFull />,
       },
     ],
   },
