+++ src/components/AudioPlayerComponents/AudioPlayerContainer/AudioPlayerContainer.tsx
@@ -13,14 +13,25 @@ const AudioPlayerContainer = (/*props: AudioPlayerContainerProps*/) => {
   // const { width, breakpoint } = props;
 
   return (
-    <div className="flex items-center justify-between w-full mt-12 border">
-      {/* {width > breakpoint && (
+    <div
+      className="mt-12"
+      style={{
+        position: 'fixed',
+        bottom: 0,
+        left: 0,
+        right: 0,
+        backgroundColor: 'rgba(0, 180, 249)',
+      }}
+    >
+      <div className="flex items-center justify-between w-full border mt-2">
+        {/* {width > breakpoint && (
         <img className="h-[50px] mx-[50px]" src={tanpuraLeft} alt="tanpura" />
       )} */}
-      <AudioPlayer />
-      {/* {width > breakpoint && (
+        <AudioPlayer />
+        {/* {width > breakpoint && (
         <img className="h-[50px] mx-[50px]" src={tanpuraRight} alt="tanpura" />
       )} */}
+      </div>
     </div>
   );
 };
