+++ src/views/tributes/Tributes.tsx
@@ -1,17 +1,17 @@
 import styled from '@emotion/styled';
 
-import { TestimonialCard } from './TestimonialCard';
+import { TributeCard } from './TributeCard';
 
-type TestimonialType = 'text' | 'video';
+type TributeType = 'text' | 'video';
 
-export interface Testimonial {
+export interface Tribute {
   id: number;
   submittedBy: string;
-  type: TestimonialType;
+  type: TributeType;
   content: string;
 }
 
-const testimonials: Testimonial[] = [
+const tributes: Tribute[] = [
   {
     id: 1,
     submittedBy: 'Obcaecati Fuga',
@@ -74,13 +74,13 @@ const testimonials: Testimonial[] = [
   },
 ];
 
-const TestimonialsWrapper = styled.div();
-const TestimonialsList = styled.div();
+const TributesWrapper = styled.div();
+const TributesList = styled.div();
 
-export function Testimonials() {
+export function Tributes() {
   return (
-    <TestimonialsWrapper className="mx-8 text-center">
-      <h1>Testimonials</h1>
+    <TributesWrapper className="mx-8 text-center">
+      <h1>Tributes</h1>
       <p>
         Acharya Roop Verma touched the lives of countless individuals over the ocurse of
         his life. This page is a modest attempt to collect the testimonimals from some of
@@ -88,11 +88,11 @@ export function Testimonials() {
         to SOMEMAIL@EMAIL.EMAIL
       </p>
       {/** breakpoint to justify-center at smaller screensizes */}
-      <TestimonialsList className="flex flex-wrap justify-start items-center mx-auto mt-8 gap-2">
-        {testimonials.map((testimonial) => (
-          <TestimonialCard key={testimonial.id} testimonial={testimonial} />
+      <TributesList className="flex flex-wrap justify-start items-center mx-auto mt-8 gap-2">
+        {tributes.map((tribute) => (
+          <TributeCard key={tribute.id} tribute={tribute} />
         ))}
-      </TestimonialsList>
-    </TestimonialsWrapper>
+      </TributesList>
+    </TributesWrapper>
   );
 }
